'''SegmentContainerOperation'''
 * impl::Segment and Segment classes
 * SegmentContainer class
 * SegmentContainerOperator class
 * New file src/Function/Operator/SegmentContainerOperator.h
 * Remove src/Data/Container/SegmentedContainer.h

+++ projects/Proteins/Data/Formats/PSSMFileParser.h
@@ -103,7 +103,7 @@ public:
     String gapScore = line.substring(153, 157).trim();
     std::cout << gapScore << std::endl;
     
-    scores->setProbability(aminoAcidTypeEnumeration->getNumElements(), normalize(gapScore.getDoubleValue()));
+    scores->setProbability(aminoAcidTypeEnumeration->getNumElements(), normalize(gapScore.getIntValue()));
     pssm->setElement(currentPosition, scores);
 
     ++currentPosition;
